Implement IXmlRepository storing documents in Google Cloud Storage
This has no *unit* tests as the integration tests are much more useful. Suggestions for unit tests welcome, of course.

It's not clear how ASP.NET Core supports the case where multiple servers add an element for the same purpose at the same time, but that's not something this code can affect.